Implements Serializable

+++ src/main/java/com/roadtaxapi/roadtax/OwnerRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource(collectionResourceRel = "owner", path="owner")
-public interface OwnerRepository extends JpaRepository<Owner, Integer>
+public interface OwnerRepository extends JpaRepository<Owner, Long>
 {
 }
