+++ src/main/java/com/roadtaxapi/roadtax/TaxPaymentRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource(collectionResourceRel = "taxPayment", path="taxPayment")
-public interface TaxPaymentRepository extends JpaRepository<TaxPayment, Integer>
+public interface TaxPaymentRepository extends JpaRepository<TaxPayment, Long>
 {
 }
