+++ src/main/java/com/roadtaxapi/roadtax/VehicleCategoryRelationRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource(collectionResourceRel = "vehicleCategory", path="vehicleCategory")
-public interface VehicleCategoryRelationRepository extends JpaRepository<VehicleCategoryRelation, Integer>
+public interface VehicleCategoryRelationRepository extends JpaRepository<VehicleCategoryRelation, Long>
 {
 }
