+++ src/main/java/com/roadtaxapi/roadtax/VehicleCategoryRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource(collectionResourceRel = "category", path="category")
-public interface VehicleCategoryRepository extends JpaRepository<VehicleCategory, Integer>
+public interface VehicleCategoryRepository extends JpaRepository<VehicleCategory, Long>
 {
 }
