+++ src/main/java/com/roadtaxapi/roadtax/VehicleRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource(collectionResourceRel = "vehicle", path="vehicle")
-public interface VehicleRepository extends JpaRepository<Vehicle, Integer>
+public interface VehicleRepository extends JpaRepository<Vehicle, Long>
 {
 }
